# server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DroppedNeedle v3",
    "description": "Native API contract. Generated clients must never be hand-edited.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v3/admin/users": {
      "get": {
        "tags": [
          "crate::auth::users::handlers"
        ],
        "summary": "Admin user listing, paged.",
        "operationId": "admin_list_users",
        "responses": {
          "200": {
            "description": "User page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::auth::users::handlers"
        ],
        "summary": "Create a local account.",
        "operationId": "admin_create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/auth/login": {
      "post": {
        "tags": [
          "crate::auth::routes::native"
        ],
        "summary": "Local login. Cookie mode (default) sets the session cookie and the body\ncarries no token; Bearer mode returns the raw token once and sets no\ncookie. Failures are one uniform 401 with a dummy-hash verify behind it.",
        "operationId": "login_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authenticated user; token only in Bearer mode",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Login/setup success body: the authenticated user, plus the raw session\ntoken only in Bearer mode (cookie mode carries it in `Set-Cookie`).",
                  "required": [
                    "user"
                  ],
                  "properties": {
                    "token": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "Raw session token; present only in Bearer mode."
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse",
                      "description": "Authenticated user."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid username or password"
          }
        }
      }
    },
...
    "/api/v3/auth/setup": {
      "post": {
        "tags": [
          "crate::auth::routes::native"
        ],
        "summary": "First-admin setup. Creates the admin and logs them in (201 + session,\nv2 parity) iff no users exist; otherwise 409. Concurrent setups serialize\non the state guard: the loser probes after the winner commits and lands\non 409. (The users slice has no empty-table constraint, so without the\nguard two distinct usernames would both succeed.)",
        "operationId": "setup_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "First admin created and logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Login/setup success body: the authenticated user, plus the raw session\ntoken only in Bearer mode (cookie mode carries it in `Set-Cookie`).",
                  "required": [
                    "user"
                  ],
                  "properties": {
                    "token": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "Raw session token; present only in Bearer mode."
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse",
                      "description": "Authenticated user."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid username, password, or email"
          },
          "409": {
            "description": "Setup has already been completed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUserCreate": {
        "type": "object",
        "description": "Admin user creation.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name; defaults to the username casing."
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional email."
          },
          "password": {
            "type": "string",
            "description": "Initial password, at least 12 chars."
          },
          "role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Role",
                "description": "Defaults to `user`."
              }
            ]
          },
          "username": {
            "type": "string",
            "description": "Login username, 3-32 chars."
          }
        }
      },
      "AdminUserListResponse": {
        "type": "object",
        "description": "Admin user listing.",
        "required": [
          "users",
          "total"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total user count.",
            "minimum": 0
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "description": "One page of users."
          }
        }
      },
...
      "LoginBody": {
        "type": "object",
        "description": "Local login body.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "Account password."
          },
          "transport": {
            "$ref": "#/components/schemas/TransportDto",
            "description": "Credential handoff; defaults to cookie."
          },
          "username": {
            "type": "string",
            "description": "Username, matched case-insensitively."
          }
        }
      },
...
      "Role": {
        "type": "string",
        "description": "Account role. Meanings are unchanged from v2: `user` requests await\napproval; `trusted` and `admin` auto-approve and are quota-exempt.",
        "enum": [
          "user",
          "trusted",
          "admin"
        ]
      },
...
      "SetupBody": {
        "type": "object",
        "description": "First-admin setup body. Runs only against an empty user table.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name; defaults to the username casing."
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional email."
          },
          "password": {
            "type": "string",
            "description": "Initial password, at least 12 chars."
          },
          "transport": {
            "$ref": "#/components/schemas/TransportDto",
            "description": "Credential handoff; defaults to cookie."
          },
          "username": {
            "type": "string",
            "description": "Login username, 3-32 chars."
          }
        }
      },
...
      "TransportDto": {
        "type": "string",
        "description": "Credential handoff for session-issuing routes. Mirrors the sibling\n`TransportParam` (which owns the mechanism) so this layer stays the only\nplace that needs `ToSchema`.",
        "enum": [
          "cookie",
          "bearer"
        ]
      },
      "UserResponse": {
        "type": "object",
        "description": "Public account shape.",
        "required": [
          "id",
          "display_name",
          "role",
          "providers",
          "created_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Served avatar URL, when an avatar is set."
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation time, unix seconds."
          },
          "display_name": {
            "type": "string",
            "description": "Display name."
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "Email, when set."
          },
          "id": {
            "type": "string",
            "description": "User id."
          },
          "last_login_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Last login time, unix seconds."
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bound provider names (e.g. `local`)."
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Account role."
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "description": "Lowercased login identifier, when set."
          },
          "username_display": {
            "type": [
              "string",
              "null"
            ],
            "description": "Preferred username casing, when set."
          }
        }
      },
...
    }
  }
}